requirements.txt
langchain
langchain-ollama
langchain-pinecone
pinecone-client
python-dotenv
flask
requests
